requirements.txt
absl-py==2.2.0
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
apache-tvm==0.14.dev273
asttokens==3.0.0
attrs==24.3.0
bitblas
Bottleneck==1.4.2
chardet==5.2.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cpplint==2.0.2
cycler==0.12.1
Cython==3.0.12
DataProperty==1.1.0
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docutils==0.21.2
dtlib==0.0.0.dev2
einops==0.8.1
evaluate==0.4.3
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
flash_attn==2.7.4.post1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.3.1
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.2.1
gradio==3.40.0
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
hf-xet==1.1.0
httpcore==1.0.7
httpx==0.25.2
huggingface-hub==0.31.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.1.0
intel-cmplr-lib-ur==2025.0.5
intel-openmp==2025.0.5
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
latex2mathml==3.77.0
linkify-it-py==2.0.3
lxml==5.3.1
markdown-it-py==2.2.0
markdown2==2.5.3
MarkupSafe==2.1.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl==2025.0.1
mkl-service==2.4.0
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.15
narwhals==1.38.2
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.20
nltk==3.9.1
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
openai==1.75.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathvalidate==3.2.3
patsy==1.0.1
peft==0.14.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
portalocker==3.1.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pybind11==2.13.6
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pytablewriter==1.2.1
pytest==8.3.5
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
PyYAML
pyzmq==26.2.0
RapidFuzz==3.13.0
ray==2.42.0
referencing==0.36.2
regex==2024.11.6
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
ruff==0.11.2
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
shellingham==1.5.4
shortuuid==1.0.13
smmap==5.0.2
sniffio==1.3.1
sqlitedict==2.1.0
stack-data==0.6.3
starlette==0.45.3
svgwrite==1.4.3
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tbb==2022.0.0
tcmlib==1.2.0
tcolorpy==0.1.7
termcolor==3.0.1
thefuzz==0.22.1
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.5
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.51.3
triton==3.1.0
typepy==1.3.4
typer==0.15.2
tzdata==2023.3
uc-micro-py==1.0.3
umf==0.9.1
uvicorn==0.34.0
wandb==0.19.11
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
word2number==1.1
xxhash==2.0.2
yarl==1.18.0
zipp==3.21.0
zstandard==0.23.0

